+++ test/query.test.js
@@ -8,13 +8,10 @@ let QueryModel;
 // Wait for the database
 beforeAll((done) => {
 
-    // Just in case the database connection takes a while.
-    jest.setTimeout(10000);
-
     // Init Linz.
     linz.init({
         options: {
-            'mongo': 'mongodb://localhost:27777/query-test',
+            'mongo': 'mongodb://mongodb:27017/query-test',
             'user model': 'user',
             'load models': false,
             'load configs': false
@@ -92,7 +89,7 @@ beforeAll((done) => {
 
     });
 
-});
+}, 10000);
 
 afterAll((done) => linz.mongoose.connection.close(done));